LPS-57626 Migrate and fix java Render/Action
