+++ modules/apps/login/login-web/src/com/liferay/login/web/portlet/action/CreateAccountMVCRenderCommand.java
@@ -12,32 +12,30 @@
  * details.
  */
 
-package com.liferay.portlet.login.action;
+package com.liferay.login.web.portlet.action;
 
+import com.liferay.login.web.constants.LoginPortletKeys;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
-import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
-import com.liferay.portal.kernel.util.JavaConstants;
-import com.liferay.portal.model.Company;
 import com.liferay.portal.theme.ThemeDisplay;
-import com.liferay.portal.util.PortletKeys;
 import com.liferay.portal.util.WebKeys;
 
-import javax.portlet.PortletConfig;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
+
+import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Peter Fellwock
  */
-@OSGiBeanProperties(
+@Component(
 	property = {
-		"javax.portlet.name=" + PortletKeys.FAST_LOGIN,
-		"javax.portlet.name=" + PortletKeys.LOGIN,
-		"mvc.command.name=/login/create_anonymous_account"
-	}
+		"javax.portlet.name=" + LoginPortletKeys.FAST_LOGIN,
+		"javax.portlet.name=" + LoginPortletKeys.LOGIN,
+		"mvc.command.name=/login/create_account"
+	},
+	service = MVCRenderCommand.class
 )
-public class CreateAnonymousAccountMVCRenderCommand
-	implements MVCRenderCommand {
+public class CreateAccountMVCRenderCommand implements MVCRenderCommand {
 
 	@Override
 	public String render(
@@ -46,24 +44,9 @@ public class CreateAnonymousAccountMVCRenderCommand
 		ThemeDisplay themeDisplay = (ThemeDisplay)renderRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		Company company = themeDisplay.getCompany();
+		renderResponse.setTitle(themeDisplay.translate("create-account"));
 
-		if (!company.isStrangers()) {
-			return "/html/portlet/login/login.jsp";
-		}
-
-		PortletConfig portletConfig = (PortletConfig)renderRequest.getAttribute(
-			JavaConstants.JAVAX_PORTLET_CONFIG);
-
-		String portletName = portletConfig.getPortletName();
-
-		if (!portletName.equals(PortletKeys.FAST_LOGIN)) {
-			return "/html/portlet/login/login.jsp";
-		}
-
-		renderResponse.setTitle(themeDisplay.translate("anonymous-account"));
-
-		return "/html/portlet/login/create_anonymous_account.jsp";
+		return "/create_account.jsp";
 	}
 
 }
